+++ extensions/css-language-features/package.json
@@ -23,9 +23,6 @@
       "supported": true
     }
   },
-  "enabledApiProposals": [
-    "documentPaste"
-  ],
   "scripts": {
     "compile": "npx gulp compile-extension:css-language-features-client compile-extension:css-language-features-server",
     "watch": "npx gulp watch-extension:css-language-features-client watch-extension:css-language-features-server",
