+++ extensions/vscode-api-tests/package.json
@@ -17,7 +17,6 @@
     "defaultChatParticipant",
     "diffCommand",
     "documentFiltersExclusive",
-    "documentPaste",
     "editorInsets",
     "embeddings",
     "extensionRuntime",
